Add Danger as our first always-on discoverer
