+++ src/build-discoverers/npm.js
@@ -5,27 +5,31 @@ import BuildDiscoverBase from '../build-discover-base';
 
 const d = require('debug')('surf:build-discover-npm');
 
-export default class DangerBuildDiscoverer extends BuildDiscoverBase {
+export default class NpmBuildDiscoverer extends BuildDiscoverBase {
   constructor(rootDir) {
     super(rootDir);
-    
-    // Danger runs concurrently with other builds
-    this.shouldAlwaysRun = true;
   }
 
   async getAffinityForRootDir() {
-    let dangerFile = path.join(this.rootDir, 'Dangerfile');
-    let exists = await statNoException(dangerFile);
+    let pkgJson = path.join(this.rootDir, 'package.json');
+    let exists = await statNoException(pkgJson);
     
-    if (exists) { d(`Found Dangerfile at ${dangerFile}`); }
-    return exists ? 2 : 0;
+    if (exists) { d(`Found package.json at ${pkgJson}`); }
+    return exists ? 5 : 0;
   }
 
   async getBuildCommand() {
+    let pkgJson = JSON.parse(
+      await fs.readFile(path.join(this.rootDir, 'package.json'), 'utf8'));
+      
     let cmds = [
-      { cmd: 'bundle', args: ['exec', 'danger']}
+      { cmd: 'npm', args: ['install']}
     ];
-  
+    
+    if (pkgJson.scripts && pkgJson.scripts.test) {
+      cmds.push({ cmd: 'npm', args: ['test']});
+    }
+    
     return {cmds};
   }
 }
